Moving to super project.

+++ src/maidsafe/routing/bootstrap_file_handler.cc
@@ -43,7 +43,7 @@ std::vector<boost::asio::ip::udp::endpoint>
   boost::asio::ip::udp::endpoint endpoint;
   for (int i = 0; i < protobuf_bootstrap.bootstrap_contacts().size(); ++i) {
     endpoint.address().from_string(protobuf_bootstrap.bootstrap_contacts(i).ip());
-    endpoint.port(protobuf_bootstrap.bootstrap_contacts(i).port());
+    endpoint.port(static_cast<unsigned short>(protobuf_bootstrap.bootstrap_contacts(i).port()));
     bootstrap_nodes[i] = endpoint;
   }
   return  bootstrap_nodes;
